requirements.txt
python-frontmatter==0.5.0
Markdown==3.2.1
Jinja2==2.11.2
awscli==1.18.41
mdx-linkify==1.4